Extract squareBtn into reusable AppBarSquareButton widget
Moves the inline squareBtn helper from FullScreenMenuPage._TopBar into a
new shared widget at lib/widgets/app_bar_square_button.dart, then reuses
it in TripsPage._PageHeader to replace the duplicate _HeaderIconButton and
_AdaptiveFilterButton private classes.
